backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]
alembic
psycopg2-binary
pydantic_settings 
email-validator
